fix #1240, closes #1241, closes #1242, attempt pypi load
