+++ src/ch01_allot/__init__.py
@@ -1,0 +1,1 @@
+pass
