+++ src/ch02_contact/__init__.py
@@ -1,0 +1,1 @@
+pass
