+++ src/ch04_rope/__init__.py
@@ -1,0 +1,1 @@
+pass
