+++ src/ch06_plan/__init__.py
@@ -1,0 +1,1 @@
+pass
