+++ src/ch09_person_lesson/__init__.py
@@ -1,0 +1,1 @@
+pass
