+++ src/ch10_person_listen/__init__.py
@@ -1,0 +1,1 @@
+pass
