+++ src/ch27_lego/__init__.py
@@ -1,0 +1,1 @@
+pass
