+++ src/ch32_world/__init__.py
@@ -1,0 +1,1 @@
+pass
