+++ src/ch34_finance/__init__.py
@@ -1,0 +1,1 @@
+pass
