+++ src/ch35_person_viewer/static/__init__.py
@@ -1,0 +1,1 @@
+pass
