+++ src/ch98_linter/__init__.py
@@ -1,0 +1,1 @@
+pass
